admin control, add/remove categories (cant delete ones in use). user delete/ toggle block. users get your account is blocked msg when logging in

+++ backend/controllers/categoryController.js
@@ -1,5 +1,6 @@
 const asyncHandler=require('express-async-handler')
 const Category=require('../models/categoryModel')
+const Post=require('../models/postModel')
 
 
 const addCategory=asyncHandler(async(req,res)=>{
@@ -16,10 +17,16 @@ const addCategory=asyncHandler(async(req,res)=>{
 
 const deleteCategory=asyncHandler(async(req,res)=>{
     const category=await Category.findById(req.params.id)
+    const categoryInUse=await Post.exists({category:req.params.id})
     if(!category){
         res.status(400)
         throw new Error('Category not found')
     }
+    if(categoryInUse){
+        res.status(400)
+        throw new Error('Category in use!')
+    }
+
     await Category.findByIdAndDelete(req.params.id)
     res.status(200).json({id:req.params.id})
 })
@@ -31,7 +38,6 @@ const getCategories=asyncHandler(async(req,res)=>{
         res.status(400)
         throw new Error('No categories')
     }
-    console.log(categories)
     res.status(200).json(categories)
 })
 module.exports={addCategory,deleteCategory,getCategories}
